+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* Changed the retry behavior for operation error `TIMEOUT` from `NonRetryable` to `ConditionallyRetryable`
+
 ## v3.136.0
 * The `query.WithResponsePartPrefetch(n)` method has been added to enable the prefetching of parts of query results.
   By default, this feature is disabled. Prefetching produces the following effects: